requirements.txt
# Core FastAPI dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# LangChain and AI
langchain
langgraph
pydantic-settings
langchain-groq==0.0.1
langchain-mcp-adapters==0.1.0

# Google API libraries
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.115.0

# MCP (Model Context Protocol)
mcp==0.1.0

# Async support
aiofiles==23.2.1
httpx==0.26.0

# Testing (optional but recommended)
pytest==7.4.4
pytest-asyncio==0.23.3

# Development tools (optional)
black==24.1.1  # Code formatter
flake8==7.0.0  # Linter
mypy==1.8.0    # Type checker


faiss-cpu
numpy
PyPDF2
neo4j